+++ CMakeLists.txt
@@ -69,7 +69,6 @@ if (CAFFEINE_ENABLE_BUILD)
 
   add_subdirectory(src)
   add_subdirectory(libraries)
-  add_subdirectory(distributed)
 
   if (CAFFEINE_ENABLE_TESTS)
     add_subdirectory(test)
