+++ cmake/CaffeineFormatting.cmake
@@ -2,7 +2,7 @@
 function(create_formatting_targets)
 
 # Include any source directories that contain formattable code here
-set(checked_dirs include src test distributed)
+set(checked_dirs include src test)
 set(fmt_sources "")
 
 foreach(directory ${checked_dirs})
