--- distributed/Protos/operation.capnp
@@ -1,1 +1,0 @@
-@0xf54292a58fc34e9a
