--- distributed/src/Serialization/ValueIdMap.cpp
@@ -1,5 +1,0 @@
-#include "caffeine_distributed/Serialization/ValueIdMap.h"
-
-namespace caffeine_distributed {
-ValueIdMap::ValueIdMap(llvm::Function* func) : func_{func} {};
-}
