+++ src/CMakeLists.txt
@@ -35,7 +35,23 @@ file(
   "${CMAKE_SOURCE_DIR}/include/*.inl"
 )
 
+file(
+  GLOB capnp_schemas
+  CONFIGURE_DEPENDS
+  RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}"
+  "${CMAKE_CURRENT_SOURCE_DIR}/Protos/*.capnp"
+)
+
+
+set(CAPNPC_OUTPUT_DIR "${CMAKE_CURRENT_BINARY_DIR}/caffeine/")
+file(MAKE_DIRECTORY ${CAPNPC_OUTPUT_DIR})
+
+capnp_generate_cpp(CAPNP_SRCS CAPNP_HDRS "${capnp_schemas}")
+list(APPEND sources ${CAPNP_SRCS})
+
 add_library(caffeine STATIC ${sources} ${headers})
+
+target_include_directories(caffeine PRIVATE "${CMAKE_CURRENT_BINARY_DIR}")
 
 target_include_directories(caffeine
   PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}"
@@ -51,4 +67,4 @@ target_include_directories(caffeine SYSTEM
 )
 
 target_link_options(caffeine PUBLIC ${LINK_FLAGS})
-target_link_libraries(caffeine PUBLIC LLVMCore "${Z3_LIBRARIES}" fmt::fmt magic_enum)
+target_link_libraries(caffeine PUBLIC LLVMCore "${Z3_LIBRARIES}" fmt::fmt magic_enum CapnProto::capnp-rpc)
