+++ src/Protos/operation.capnp
@@ -1,0 +1,9 @@
+@0xf54292a58fc34e9a;
+
+
+struct Symbol {
+    union {
+        name @0 :Text;
+        number @1 :UInt64;
+    }
+}
